Add Highlighter interface
