+++ packages/guides-code/resources/config/guides-code.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Highlight\Highlighter as HighlightPHP;
 use phpDocumentor\Guides\Code\Highlighter\Highlighter;
+use phpDocumentor\Guides\Code\Highlighter\HighlightPhpHighlighter;
 use phpDocumentor\Guides\Code\Twig\CodeExtension as TwigExtension;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
@@ -15,10 +16,11 @@ return static function (ContainerConfigurator $container): void {
 
         ->set(HighlightPHP::class)
 
-        ->set(Highlighter::class)
+        ->set(HighlightPhpHighlighter::class)
         ->args([
             '$languageAliases' => [],
         ])
+        ->alias(Highlighter::class, HighlightPhpHighlighter::class)
 
         ->set(TwigExtension::class)
         ->tag('twig.extension');
